modrm: don't error out when checking if bp is base for mod=11 or no mod/rm.
When getting the segment we need to know if bp is being used as base.
Some instructions like movs need to check for the segment override and
don't have a modrm byte, so record whether there is a mod/rm and if
there isn't then default to DS/SS as appropriate.