rtl: implement prefetcher.
The prefetcher is responsible for reading the instruction stream from
memory and putting it into to a fifo ready for decode.  Fetches will
happen when there is sufficient space in the fifo.  There is no
pipelining of reads, as an unaligned transfer followed by a pipelined
read could end up overflowing the fifo and would need additional
buffering.