tests/rtl: support events.
Rather than have each test manually keep track of time and know when to
change signals, add support for events that can be deferred or run on a
clock edge.  Use posedge events for reading the state of a simulation
and negedge events for setting new variables.  Deferred events can be
used for things like emulating memories that take cycles to propogate.