tests/VerilogTestbench: rework events processing.
Synchronising combinatorial and sequential logic between the testbench
and the RTL is tricky.  Prior to this commit, the design could be
evaluated before the clock edge and so data intended for the next cycle
would look like it had been delivered before the edge rather than on the
edge, giving incorrect simulation results.  So make sure that setup
happens on or just after the clock edge and capture at the end of the
cycle.

In tests this means that rather than just setting some signals, calling
cycle() and reading them, it is cleaner to set the signals on the next
edge, have them set for a full cycle and then cycle again to trigger the
next edge.  When viewing traces this makes it much clearer what is going
on.