rtl: add the ALU.
The ALU performs all of the arithmetic, bitwise operations along with
some simple muxing.

The Q bus is tapped off before the ALU so that the MAR/MDR can be fed
back into the A/B busses, but without introducing a combinational loop
through A/B into the ALU and back out.  Therefore it is not legal to
have A/B driven from Q and for Q to be driven from the ALU.