tests/core: get instruction length with a DPI call.
The previous method of examining changes in the instruction pointer is
fundamentally broken in the case of jumps.  Instead, use a DPI call to
count the number of FIFO pops.  This is unfortunate in that it means it
won't work on hardware, but it's fine for simulation.

Hardware will likely need to skip the instruction length checks.