rtl/core: synchronize CS:IP updates to prefetcher.
Synchronize CS and IP updates so that they can be deployed to the
prefetcher atomically.  For something like a far jump, the microcode
would first update the IP and then update CS.  If this happens during
the microcode execution then potentially the prefetcher can start
prefetching from an invalid address - CS remains the same but IP changes
and we fetch from invalid/unmapped memory.

This simple module just defers the output until propagation is ready, in
this case, on completion of the microinstruction when CS:IP can be
delivered to the prefetcher in a single cycle.