microcode: allow specifying an immediate in the microcode.
There are cases where we want to pass an immediate to the ALU that isn't
in the instruction stream.  For example, when performing an indirect
inter-segment jump, we need to read the IP from the memory operand, and
then the CS at the next two bytes.  So provide a way for the immediate
to be specified by the microcode.  Omitting this field will cause the
immediate to be multiplexed from the immediate reader.

Rather than encoding all 16 bits in the microinstruction, encode an
index into a lookup table - there will be at most a handful of
immediates required.