rtl/microcode: don't abuse segment_override.
segment_override is for indicating that the opcode is a segment override
and that the core should record the override.  segment_force is subtly
different in that it forces that segment only for that microinstruction,
and is important for things like a push SR or any stack access where we
always need to use SS rather than any override segment.