rtl/sim: support for NMI+INT.
This adds support for NMIs and external interrupts.  This does deviate
slightly from the original 8086 in that interrupts correctly handle
prefix bytes during string operations.  On the 8086, if a string
instruction had multiple prefixes (say a rep prefix and a segment
override), then an interrupt during that instruction would leave IP
pointing to the middle of the prefixes so would not resume correctly.

This is correctly handled here, all prefixes are treated as part of the
instruction and so the saved IP will point to the beginning of the
prefix bytes.

There's a simple IRQ controller for the FPGA testing that isn't yet
hooked up to any peripherals.