rtl/Fifo: convert to an acked-push.
This changes the FIFO to always present the head data until acked rather
than reading on request.  This cuts out one cycle of latency when
reading and means that the opcode_dispatch microcode can be bypassed
when the FIFO is not empty, giving a fairly large performance boost -
previously a single byte instruction like "nop (0x90)" would spend two
cycles in the opcode dispatch and then one in the 0x90 opcode, this is
now a single cycle on the opcode.  This also speeds up mod/rm decode and
immediate fetching.