rtl/MemArbiter: workaround invalid synthesis.
The data access trigger for the arbiter is gated with ~d_io so that I/O
accesses don't go through the arbiter (there is a separate IO bus), but
when synthesized with Quartus, this ends up generating spurious
accesses, and the data_m_access signal also needs to be gated with ~d_io
*inside* the module.

This does not reproduce in simulation, and even stranger, the problem
occurs when running tests that do not perform any I/O at all.  It ends
up corrupting the instruction stream and executing invalid instructions.