core: implement ENTER instruction.
This is the last 80186 instruction and is a horrible one as the size
which is used last, comes before the nest level in the encoding.  To
implement the loop we introduce a simple loop counter, and also need a
specialized ALU operation for working out where to store the frame_temp
value as it's derived from the SP and nest count and the TEMP register
is used for holding the size.