bios: rework timer interrupt.
The timer interrupt was switching stacks like a normal BIOS call, but
this was a problem when running qbasic, because it looks like the
application was hooking the interrupt itself and then making a nested
BIOS call which would result in clobbering an existing stack.

Rather than supporting nested interrupts, just use the existing stack
for the timer interrupt - it only requires a couple of pushes.