ps2: add a PS/2 controller.
This is a simple device->host PS/2 keyboard controller.  host->device
communications are missing at the moment and this is a problem for some
keyboards that don't automatically come out of reset.  On these
keyboards they keep sending an invalid byte on reset, waiting for
incorrect parity detection.

A future update will add host->device communications to reset these
keyboards.