sim: add serialization support.
A release build of the SoftwareCPU simulation can run at around 10MHz on
my laptop whereas a debug build of RTLCPU writing a VCD dump may only
run at around 60Khz.  It's prohibitively expensive to boot a system in
the latter mode, but by serializing the simulator state and restoring
it, we can get to the target state with the SoftwareCPU and then load it
back into the simulator in a different mode.

The serialization does everything through the debug interface, so
serialization of the internals of the CPU does not happen, only things
that can be set with a debugger.  This means:

  - memory contents
  - general purpose registers, flags, segment registers, IP
  - peripheral registers and state

Some things are not serialized, for example the disk image, SDL window
etc as those things are regularly refreshed based on the values of
registers and so will be recomputed at the next IO processing interval.