sim: add trace for executed instructions.
Adding --trace will cause the CS:IP and instruction bytes to be written
to stdout.  scripts/disassemble-trace can then be used to disassemble
that at a later time for diagnostics.

Note: this will not do the right thing if an instruction overwrites
itself - the instruction is read once the length is known.