rtl/Fifo: don't use asynchronous reset for flushing.
This is a particularly nasty one that lay hidden for a long time,
masquerading as a timing problem with the SDRAM.  The prefetch FIFO
should be reset under two conditions - global reset or change in CS:IP.
The global reset case is an edge triggered active high reset and this
affects the entire core without problems, but for the change in CS:IP,
this is something that happens during normal operation.  The problem was
that the fifo reset was derived from detecting a write to CS:

    wire cs_updating = seg_wr_sel == CS && segment_wr_en;

and this is problematic becuase seg_wr_sel is a 2-bit bus.  During a
microinstruction that writes to DS (2'b11) rather than CS (2'b01),
different propagation delays between the two bits could end up
triggering an edge on cs_updating and falsely resetting the FIFO.  The
particularly nasty thing about this is that it's highly dependent on
place-and-route, and in some cases, only triggered a reset of part of
the FIFO - often the read/write pointers would be unchanged but the
count would reset itself to 0.

The net result was that the FIFO could end up with corrupted memory
because the empty/full logic was wrong and then the core would execute
incorrect instructions.  This doesn't manifest in simulators where those
delays between the seg_wr_sel wires weren't modelled and so would never
show up in simulation.

The simple fix here is to split the reset into a global reset and a
flush operation.  The global reset is edge triggered, the flush
operation is sampled on the clock posedge.