rtl: add a direct mapped cache.
This is a direct mapped writeback cache with a 16 byte line size.  The
critical word is filled first.

The cache does not reset because the CPU is not cache aware and expects
that on reset the memory retains its contents.  The cache is disabled on
the simulator to maximise performance (where the latency is set to 0)
and because there is no separate video memory.

This gives a 2X performance boost in topbench and 3x boost in landmark.