rtl/Core: add a hardware instruction decoder.
This large commit replaces the microcode based instruction decoding with
a hardware decoder.  The core now has a loosely coupled 3 stage pipeline
consisting of fetch, decode and execute.  Whilst this doesn't offer a
noticable performance increase in microbenchmarks, it does reduce the
latency of many instructions and also has fully decoded instructions at
the start of each microprogram which is a prerequisite for protected
mode where segment limits need to be verified early on.

One departure from the 80186 is that the hardware decoding means there
is a limit on instruction length.  Currently any instruction of 256
bytes or more which is not a valid encoding but possible with redundant
prefix bytes will result in undefined behaviour.