core: add a pseudo-286 mode.
Most 286 real mode applications don't use any of the protected mode
features or exceptions and really just want the new instructions that
are also present in the 80186 such as immediate shifts etc.  Add a
config option that allows the core to identify as an 80286 to let those
applications run.  This works for Monkey Island 2: LeChuck's Revenge and
Indiana Jones and the Fate of Atlantis at least.