bios/sd: discard initial ff response bytes
When receiving a response, particularly for a read/write there can be a
large number of FF bytes sent back from the card.  This is a problem
because for slow cards we can overflow the buffer, and when looking for
responses we have more bytes to scan through.  So just discard those
useless bytes.