bios: don't enable interrupts in bios calls.
The BIOS isn't re-entrant so can't safely take hardware interrupts that
call an application defined hook such as the mouse or a timer that might
then call back into the bios.  Instead, disable the mouse during
long-running SD accesses that might result in losing synchronization
with the mouse and try to resynchronize when possible.